Carried out the tasks from the test
This code is now capable of grabbing data from the restful API, handle the json and use it to populate a table as well as generating users age and age in 20 years and gathering the most popular colour and displaying that on a table sorting by quantity then Alphabetically. It can then post this data back to the restful API as json successfully.